Added: Test prelude for color565 splitting
